+++ cardano-ping/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 0.2.0.4 -- 2023-06-12
 
+* Using `ISO8601` time format.
 * Only print negotiated version, if negotiation took place on the remote side.
+* Fixed formatting of ping messages.
 
 ## 0.2.0.3 -- 2023-06-09